+++ attributes/packages.rb
@@ -13,3 +13,5 @@ default['nodejs']['packages'] = value_for_platform_family(
 # Add options and actions by package name
 default['nodejs']['package_action'] = { default: :install }
 default['nodejs']['package_options'] = { default: '' }
+# Disable upstream DNF module on EL8 based systems
+default['nodejs']['dnf_module'] = false